# web/assets/openapi.yaml
openapi: 3.0.3
info:
  title: RenoP HTTP API
  version: "2.0"
  contact:
    url: https://github.com/404Setup/RenoP
  description: |
    Management and registry APIs for RenoP, a self-hosted artifact repository server
    supporting Maven, unstructured file storage, npm, Cargo (Rust sparse index), and Docker/OCI registries.

    Default listen address: `0.0.0.0:3000`.

    **Route layout:**
    - Management API: `/api/...`
    - Maven artifact storage: `/{repo}/{path}` (Maven 2 layout)
    - Unstructured file storage: `/{repo}/{path}`
    - Docker/OCI Distribution API: `/v2/...`
    - Cargo sparse registry: `/{repo}/config.json` and `/{repo}/api/v1/...`
    - npm registry: `/{repo}/{package}` and `/{repo}/-/...`
    - Javadoc preview: `/javadoc/{repo}/{path}`
    - Cargo-doc preview: `/cargodoc/{repo}/{path}`

    **Wire format:**
    Most management endpoints use `application/x-protobuf`. Schemas here describe the
    logical fields (snake_case) for documentation and client stub generation.
    See `proto/api/v1/api.proto` for canonical protobuf definitions.
    Docker/OCI, npm, and Cargo endpoints use JSON.

    **Error bodies:**
    Management errors are plain text (`Unauthorized`, `Forbidden`, `Not found`).
    Docker browser-management errors additionally expose a stable `X-Renop-Error-Code` header for localization.
    Docker Distribution API errors follow `{"errors":[{"code":"...","message":"...","detail":"..."}]}`.
...

  license:
    name: MPL-2.0
...
servers:
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: renop_session
      description: HttpOnly session cookie issued by `POST /api/auth/login`. Idle timeout is approximately 7 days.
    basicAuth:
      type: http
      scheme: basic
      description: |
        `username:password` or `username:API-token` encoded as HTTP Basic credentials.
        Basic credentials are accepted only by standard package protocols, Docker token exchange, and chunked upload.
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        A prefixed API-token secret. API endpoints intersect the token's scopes with the owning account's current
        permissions. Query-string credentials and `Authorization: Session` are rejected.
...
  responses:
...
    PlainNotFound:
      description: Resource not found. Some endpoints also return this for permission denials to avoid information disclosure.
      content:
        text/plain:
          schema:
            type: string
            example: Not found
...
  schemas:
...
    SuperTeam:
      type: object
      required: [ prefix, name, description, created_by, role_level, member_count, created_at, updated_at ]
      properties:
        prefix:
          type: string
          minLength: 2
          maxLength: 64
          pattern: "^[a-z0-9](?:[a-z0-9_-]*[a-z0-9])$"
          description: Immutable namespace shared by every supported package engine.
        name:
          type: string
          maxLength: 80
        description:
          type: string
          maxLength: 512
        created_by:
          type: string
          description: Current creator username, or the retained creation-time name after account deletion.
        role_level:
          type: integer
          minimum: 0
          maximum: 4
          description: Current account role. Zero is returned only for a system administrator who is not a member.
        member_count:
          type: integer
          minimum: 1
        created_at:
          type: integer
          format: int64
        updated_at:
          type: integer
          format: int64
...
    SuperTeamDetails:
      type: object
      required: [ team, members, administrator ]
      properties:
        team:
          $ref: "#/components/schemas/SuperTeam"
        members:
          type: array
          items:
            $ref: "#/components/schemas/SuperTeamMember"
        administrator:
          type: boolean
...
    SuperTeamConfig:
      type: object
      required: [ create_limit, join_limit ]
      properties:
        create_limit: { type: integer, minimum: 1, maximum: 1000, default: 5 }
        join_limit: { type: integer, minimum: 1, maximum: 1000, default: 20 }
...
paths:
...
  /api/super-teams:
    get:
      tags: [ super-teams ]
      summary: List visible global teams
      description: Returns the caller's teams in immutable-prefix order. System administrators see every team.
      operationId: listSuperTeams
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 12 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
      responses:
        "200":
          description: Bounded team page.
          content:
            application/json:
              schema:
                type: object
                required: [ teams, total, limit, offset, administrator ]
                properties:
                  teams: { type: array, items: { $ref: "#/components/schemas/SuperTeam" } }
                  total: { type: integer, minimum: 0 }
                  limit: { type: integer }
                  offset: { type: integer }
                  administrator: { type: boolean }
    post:
      tags: [ super-teams ]
      summary: Create a global team
      description: Reserves an immutable cross-engine prefix and assigns the caller T4 ownership.
      operationId: createSuperTeam
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ prefix, name, description ]
              properties:
                prefix: { type: string, minLength: 2, maxLength: 64 }
                name: { type: string, maxLength: 80 }
                description: { type: string, maxLength: 512 }
      responses:
        "201":
          description: Team created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuperTeam" }
        "409":
          description: Prefix already exists or the caller reached a creation/membership limit.
  /api/super-teams/eligible:
    get:
      tags: [ super-teams ]
      summary: List global teams eligible to own a resource
      description: Returns only memberships at or above the requested role; administrator status does not bypass membership.
      operationId: listEligibleSuperTeams
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      parameters:
        - { name: minimum_role, in: query, schema: { type: integer, minimum: 1, maximum: 4, default: 3 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 12 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
      responses:
        "200":
          description: Bounded eligible-team page.
          content:
            application/json:
              schema:
                type: object
                properties:
                  teams: { type: array, items: { $ref: "#/components/schemas/SuperTeam" } }
                  total: { type: integer, minimum: 0 }
                  limit: { type: integer }
                  offset: { type: integer }
  /api/super-teams/{prefix}:
    parameters:
      - name: prefix
        in: path
        required: true
        schema: { type: string, minLength: 2, maxLength: 64 }
    get:
      tags: [ super-teams ]
      summary: Get a global team
      operationId: getSuperTeam
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      responses:
        "200":
          description: Team metadata and username-only member list.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuperTeamDetails" }
        "404": { $ref: "#/components/responses/PlainNotFound" }
    put:
      tags: [ super-teams ]
      summary: Update mutable global-team metadata
      description: T4 owners and system administrators may change the name and description; the prefix is immutable.
      operationId: updateSuperTeam
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ name, description ]
              properties:
                name: { type: string, maxLength: 80 }
                description: { type: string, maxLength: 512 }
      responses:
        "200":
          description: Updated team details.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuperTeamDetails" }
    delete:
      tags: [ super-teams ]
      summary: Delete a global team
      description: T4 owners and system administrators may delete a team only when it owns no resources and has no pending transfer. Pending invitations are cancelled atomically.
      operationId: deleteSuperTeam
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      responses:
        "204": { description: Team deleted. }
  /api/super-teams/{prefix}/members:
    post:
      tags: [ super-teams ]
      summary: Invite or administratively add members
      description: T3 can invite T1/T2; T4 can invite T1-T4. System administrators add immediately and still enforce account membership limits.
      operationId: addSuperTeamMembers
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      parameters:
        - name: prefix
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ users, level ]
              properties:
                users:
                  type: array
                  minItems: 1
                  maxItems: 20
                  items: { type: string }
                level: { type: integer, minimum: 1, maximum: 4 }
      responses:
        "201": { description: Invitations sent or members added. }
        "409": { description: Existing membership/invitation or target membership limit reached. }
  /api/super-teams/{prefix}/members/{username}:
    parameters:
      - { name: prefix, in: path, required: true, schema: { type: string } }
      - { name: username, in: path, required: true, schema: { type: string } }
    put:
      tags: [ super-teams ]
      summary: Change a global-team role
      description: T3 can manage only T1/T2. T4 and system administrators may manage T1-T4. The last T4 cannot be demoted.
      operationId: updateSuperTeamMember
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ level ]
              properties:
                level: { type: integer, minimum: 1, maximum: 4 }
      responses:
        "200": { description: Role updated. }
        "409": { description: The operation would remove the last T4 owner. }
    delete:
      tags: [ super-teams ]
      summary: Remove a member or leave a global team
      operationId: removeSuperTeamMember
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      responses:
        "204": { description: Member removed. }
        "409": { description: The operation would remove the last T4 owner. }
  /api/super-teams/invitations/{id}/{decision}:
    post:
      tags: [ super-teams ]
      summary: Respond to a global-team invitation
      operationId: respondSuperTeamInvitation
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: decision, in: path, required: true, schema: { type: string, enum: [ accept, reject ] } }
      responses:
        "200": { description: Invitation answered exactly once. }
        "409": { description: Invitation expired, was already answered, or membership limit was reached. }
  /api/super-teams/limits:
    get:
      tags: [ super-teams ]
      summary: Get current account global-team limits
      operationId: getOwnSuperTeamLimits
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      responses:
        "200":
          description: Effective limits, inheritance flags, and usage.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuperTeamLimitStatus" }
  /api/super-teams/users/{username}/limits:
    parameters:
      - { name: username, in: path, required: true, schema: { type: string } }
    get:
      tags: [ super-teams ]
      summary: Get one account's effective global-team limits
      operationId: getUserSuperTeamLimits
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      responses:
        "200":
          description: Effective account limits.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuperTeamLimitStatus" }
    put:
      tags: [ super-teams ]
      summary: Override one account's global-team limits
      description: Manager-only. Use `-1` to inherit the corresponding global value; zero prevents the operation.
      operationId: updateUserSuperTeamLimits
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ create_limit, join_limit ]
              properties:
                create_limit: { type: integer, minimum: -1, maximum: 1000 }
                join_limit: { type: integer, minimum: -1, maximum: 1000 }
      responses:
        "200":
          description: Updated effective account limits.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuperTeamLimitStatus" }
  /api/settings/super-teams:
    get:
      tags: [ settings, super-teams ]
      summary: Get global team defaults
      operationId: getSuperTeamSettings
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      responses:
        "200":
          description: Global defaults.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuperTeamConfig" }
    put:
      tags: [ settings, super-teams ]
      summary: Update global team defaults
      operationId: updateSuperTeamSettings
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SuperTeamConfig" }
      responses:
        "200":
          description: Updated global defaults.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuperTeamConfig" }
...
  /api/publication-quota/super-teams/{prefix}:
    parameters:
      - { name: prefix, in: path, required: true, schema: { type: string, minLength: 2, maxLength: 64 } }
    get:
      tags: [ publication-quotas, super-teams ]
      summary: Get a global-team publication quota
      description: Team members and system managers may read this status.
      operationId: getSuperTeamPublicationQuota
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      responses:
        "200":
          description: Effective quota and current-period usage.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicationQuotaStatus" }
        "403": { $ref: "#/components/responses/PlainForbidden" }
        "404": { $ref: "#/components/responses/PlainNotFound" }
    put:
      tags: [ publication-quotas, super-teams ]
      summary: Override a global-team publication quota
      description: Manager-only. An empty object restores global defaults.
      operationId: putSuperTeamPublicationQuota
      security:
        - sessionCookie: [ ]
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PublicationQuotaOverride" }
      responses:
        "200":
          description: Updated effective quota.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicationQuotaStatus" }
        "400": { $ref: "#/components/responses/PlainBadRequest" }
        "403": { $ref: "#/components/responses/PlainForbidden" }
...